Splitted management of Keyboard in a lib file
